nila-py-chatbot/requirements.txt
PyJWT==2.7.0
PyNaCl==1.5.0
aiohttp-retry==2.8.3
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
mercantile
msgpack==1.0.5
multidict==6.0.4
py-algorand-sdk==2.3.0
pycparser==2.21
pycryptodomex==3.18.0
pytz==2023.3
requests==2.31.0
twilio==8.5.0
urllib3<2
yarl==1.9.2
qrcode==7.1.0
Pillow==9.4.0
load-dotenv==0.1.0
